requirements.txt
# Core dependencies
python-dotenv
langchain-openai
langgraph
langgraph-supervisor

# Async HTTP and networking
aiohttp
requests

# Web scraping and parsing
beautifulsoup4
lxml

# Data processing
pandas
numpy

# Financial data
yfinance

# Natural language processing
nltk

# Caching (optional)
redis[hiredis]

# Google services (optional)
googlemaps

# Search APIs (optional)
google-search-results

# Rate limiting and utilities
aiofiles
asyncio-throttle

# Development and testing
pytest
pytest-asyncio
black
isort
mypy

# Monitoring and logging
structlog
rich

# Configuration management
pydantic
pydantic-settings

# Performance monitoring
psutil
memory-profiler

# Security
cryptography

# Streamlit and UI components
streamlit
plotly
pandas

# Additional UI enhancements (optional)
streamlit-option-menu
streamlit-aggrid
streamlit-elements

# For better charts and visualizations
matplotlib
seaborn

# For data processing in UI
numpy

# For async handling in Streamlit
asyncio-throttle